% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Call_E.R
\name{Call_Exercises}
\alias{Call_Exercises}
\alias{Call_E}
\title{R FUCTION}
\usage{
Call_E(x)
}
\arguments{
\item{x}{The number selection of the exercise (between 1 and 6).}
}
\value{
The solution for the specified exercise.
}
\description{
Call function according to the number of exercises requested.
You can review pdf <Exercises.pdf> with each process and lines of code.
}
\examples{
Call_E(1)
Call_E(2)
...
Call_E(6)
}
