% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_answer.R
\name{retrieve_answers}
\alias{retrieve_answers}
\alias{retrieve_answer}
\title{R DATA ANALYSIS}
\usage{
retrieve_answer(e_num)
}
\arguments{
\item{e_num}{The number of the exercise (between 1 and 6).}
}
\value{
The solution for the specified exercise.
}
\description{
Solution of exercises to function "Call_E"
}
